rename voltage_gate_sequence -> voltage_sequence
