+++ quam_builder/architecture/quantum_dots/voltage_sequence/__init__.py
@@ -1,9 +1,9 @@
 from .sequence_state_tracker import *
 from .gate_set import *
-from .voltage_gate_sequence import *
+from .voltage_sequence import *
 
 __all__ = [
     *sequence_state_tracker.__all__,
-    *voltage_gate_sequence.__all__,
+    *voltage_sequence.__all__,
     *gate_set.__all__,
 ]
