+++ tests/voltage_gate_sequence/test_gate_set.py
@@ -1,7 +1,7 @@
 import pytest
 
 from quam.components import SingleChannel
-from quam_builder.architecture.quantum_dots.voltage_gate_sequence.gate_set import (
+from quam_builder.architecture.quantum_dots.voltage_sequence.gate_set import (
     VoltageTuningPoint,
 )
